--- a/benchmarks/NexusMods.DataModel.Benchmarks/Program.cs
+++ b/benchmarks/NexusMods.DataModel.Benchmarks/Program.cs
@@ -5,3 +5,10 @@ using NexusMods.DataModel.Benchmarks;
 using NexusMods.DataModel.Sorting;
 
 BenchmarkRunner.Run<Sorting>();
+/*
+var sorting = new Sorting();
+sorting.NumItems = 1000;
+sorting.Setup();
+sorting.Sort();
+var a = 5;
+*/