--- a/src/Games/NexusMods.Games.BethesdaGameStudios/PluginFile.cs
+++ b/src/Games/NexusMods.Games.BethesdaGameStudios/PluginFile.cs
@@ -29,33 +29,56 @@ public record PluginFile : AGeneratedFile
             .Select(f => f.File)
             .Where(f => SkyrimSpecialEdition.PluginExtensions.Contains(f.To.Extension))
             .ToArray();
-        var results = Sorter.Sort<AModFile, RelativePath, AModFile[]>(pluginFiles, 
-            i => i.To.FileName,
-            i => GenerateRules(pluginFiles, i),
+        
+        var pluginFileRuleTuples = InitialiseFileRuleTuples(pluginFiles);
+
+        var results = Sorter.Sort<ModRuleTuple, RelativePath, ModRuleTuple[]>(pluginFileRuleTuples, 
+            i => i.Mod.To.FileName,
+            i => i.Rules, //GenerateRules(pluginFiles, i),
             RelativePath.Comparer);
 
-        await stream.WriteAllLinesAsync(results.Select(i => i.To.FileName.ToString()), token: ct);
+        await stream.WriteAllLinesAsync(results.Select(i => i.Mod.To.FileName.ToString()), token: ct);
     }
 
-    private IEnumerable<ISortRule<AModFile, RelativePath>> GenerateRules(AModFile[] modFiles, AModFile aModFile)
+    private ModRuleTuple[] InitialiseFileRuleTuples(AModFile[] pluginFiles)
+    {
+        // Init Mods
+        var pluginFileRuleTuples = GC.AllocateUninitializedArray<ModRuleTuple>(pluginFiles.Length);
+        for (int x = 0; x < pluginFileRuleTuples.Length; x++)
+        {
+            ref var entry = ref pluginFileRuleTuples[x];
+            entry.Mod = pluginFiles[x];
+        }
+
+        // Generate & Cache Rules
+        for (int x = 0; x < pluginFileRuleTuples.Length; x++)
+        {
+            ref var entry = ref pluginFileRuleTuples[x];
+            entry.Rules = GenerateRules(pluginFileRuleTuples, entry.Mod).ToArray();
+        }
+
+        return pluginFileRuleTuples;
+    }
+
+    private IEnumerable<ISortRule<ModRuleTuple, RelativePath>> GenerateRules(ModRuleTuple[] modFiles, AModFile aModFile)
     {
         var defaultIdx = DefaultOrdering.IndexOf(aModFile.To.FileName);
         switch (defaultIdx)
         {
             case 0:
-                yield return new First<AModFile, RelativePath>();
+                yield return new First<ModRuleTuple, RelativePath>();
                 break;
             case > 1:
             {
                 for (var i = 0; i < defaultIdx; i++)
                 {
-                    yield return new After<AModFile, RelativePath>(DefaultOrdering[i]);
+                    yield return new After<ModRuleTuple, RelativePath>(DefaultOrdering[i]);
                 }
 
                 foreach (var itm in modFiles)
                 {
-                    if (DefaultOrdering.Contains(itm.To.FileName)) continue;
-                    yield return new Before<AModFile, RelativePath>(itm.To.FileName);
+                    if (DefaultOrdering.Contains(itm.Mod.To.FileName)) continue;
+                    yield return new Before<ModRuleTuple, RelativePath>(itm.Mod.To.FileName);
                 }
 
                 break;
@@ -66,19 +89,19 @@ public record PluginFile : AGeneratedFile
                 {
                     foreach (var dep in itm.Masters)
                     {
-                        yield return new After<AModFile, RelativePath>(dep);
+                        yield return new After<ModRuleTuple, RelativePath>(dep);
                     }
                 }
 
                 if (aModFile.To.Extension == SkyrimSpecialEdition.ESL)
                 {
-                    foreach (var file in modFiles.Where(m => m.To.Extension == SkyrimSpecialEdition.ESM))
-                        yield return new After<AModFile, RelativePath>(file.To.FileName);
+                    foreach (var file in modFiles.Where(m => m.Mod.To.Extension == SkyrimSpecialEdition.ESM))
+                        yield return new After<ModRuleTuple, RelativePath>(file.Mod.To.FileName);
                 }
                 else if (aModFile.To.Extension == SkyrimSpecialEdition.ESP)
                 {
-                    foreach (var file in modFiles.Where(m => m.To.Extension != SkyrimSpecialEdition.ESP))
-                        yield return new After<AModFile, RelativePath>(file.To.FileName);
+                    foreach (var file in modFiles.Where(m => m.Mod.To.Extension != SkyrimSpecialEdition.ESP))
+                        yield return new After<ModRuleTuple, RelativePath>(file.Mod.To.FileName);
                 }
 
 
@@ -93,5 +116,11 @@ public record PluginFile : AGeneratedFile
         await GenerateAsync(ms, loadout, flattenedList, ct);
         ms.Position = 0;
         return (Size.From(ms.Length), await ms.Hash(token: ct));
+    }
+    
+    private struct ModRuleTuple
+    {
+        public AModFile Mod;
+        public IReadOnlyList<ISortRule<ModRuleTuple, RelativePath>> Rules;
     }
 }