… benchmarks/NexusMods.DataModel.Benchmarks/Sorting.cs
… tests/NexusMods.DataModel.Tests/SortTests.cs