```csharp
| Method | NumItems |         Mean |       Error |      StdDev |      Gen0 |      Gen1 |   Allocated |
|------- |--------- |-------------:|------------:|------------:|----------:|----------:|------------:|
|   Sort |      100 |     232.5 us |     0.76 us |     0.71 us |   13.6719 |    1.9531 |   218.35 KB |
|   Sort |     1000 |   9,651.1 us |    24.59 us |    20.54 us |   46.8750 |   15.6250 |   833.33 KB |
|   Sort |     2000 |  40,248.5 us |   319.33 us |   298.70 us |   76.9231 |         - |  1696.25 KB |
|   Sort |     5000 | 197,204.1 us | 3,727.65 us | 3,661.05 us | 1000.0000 |  500.0000 | 17553.82 KB |
|   Sort |    10000 | 825,593.6 us | 6,677.17 us | 6,245.83 us | 2000.0000 | 1000.0000 | 50161.87 KB |
```

Fixes [#84.](https://github.com/Nexus-Mods/NexusMods.App/issues/84). Improves runtime by around 4x and memory usage by ~100x for big sets.
Additionally improves the benchmark to automatically benchmark against a variable number of items (from 100 to 10000).

Further improvements are still possible [we are mostly bottlenecked by lack of custom collections] but not currently worth in terms of development time.